sim-rs: handle stream processing in web worker
