+++ ui/src/components/Sim/modules/Controls.tsx
@@ -1,12 +1,10 @@
 import { FC, memo, useCallback } from "react";
 
-import { useHandlers } from "@/components/Graph/hooks/useHandlers";
 import { defaultAggregatedData, useSimContext } from "@/contexts/SimContext/context";
 import { useStreamMessagesHandler } from "../hooks/useStreamMessagesHandler";
 
 export const Controls: FC = memo(() => {
   const { state, dispatch } = useSimContext();
-  const { handleResetSim } = useHandlers();
   const { startStream, streaming, stopStream } = useStreamMessagesHandler();
 
   const handleCancelSim = useCallback(() => {
@@ -38,7 +36,7 @@ export const Controls: FC = memo(() => {
       </button>
       <button
         className="bg-gray-400 text-white w-[80px] rounded-md px-4 py-2"
-        onClick={streaming ? handleCancelSim : handleResetSim}
+        onClick={handleCancelSim}
       >
         {streaming ? "Cancel" : "Reset"}
       </button>
